+++ src/main/java/neuralnetwork/ConvolutionalNeuralNet.java
@@ -11,22 +11,22 @@ import java.util.concurrent.ThreadLocalRandom;
 public class ConvolutionalNeuralNet {//convolution > RELU > Max pooling > neuralNet
 
     //private int filterNumber;
-    private float[][][] inputs;
-    private float[][][][] filterValue;//layer,amount in layer, y, x
+    private double[][][] inputs;
+    private double[][][][] filterValue;//layer,amount in layer, y, x
     private NeuralNet fullyConnectedNLayer;
 
     public ConvolutionalNeuralNet(){
-        filterValue = new float[3][3][3][3];
+        filterValue = new double[3][3][3][3];
     }
 
     public ConvolutionalNeuralNet(int depth, int filterNumber, int filterYSize, int filterXSize, int[] neuralNetStructure){
-        filterValue = new float[depth][filterNumber][filterYSize][filterXSize];
+        filterValue = new double[depth][filterNumber][filterYSize][filterXSize];
         initRandomFilterValues();
         fullyConnectedNLayer = new NeuralNet(neuralNetStructure);
     }
 
-    public void train(float learningRate, MnistMatrix[] trainingData){
-        inputs = new float[1][][];
+    public void train(double learningRate, MnistMatrix[] trainingData){
+        inputs = new double[1][][];
         for (int i = 0; i < trainingData.length; i++) {
             inputs[0] = trainingData[i].getData();
             feedForward(inputs);
@@ -34,7 +34,7 @@ public class ConvolutionalNeuralNet {//convolution > RELU > Max pooling > neural
         }
     }
 
-    private void feedForward(float[][][] inputs){
+    private void feedForward(double[][][] inputs){
 
     }
 
@@ -42,8 +42,8 @@ public class ConvolutionalNeuralNet {//convolution > RELU > Max pooling > neural
 
     }
 
-    public float[][][] convolution(byte[][][] byteArray, int layer, int strideSize){//depth,width,lenght of data
-        float[][][] convolutionArray = new float[filterValue[layer].length][byteArray[0].length-2][byteArray[0][0].length-2];
+    public double[][][] convolution(byte[][][] byteArray, int layer, int strideSize){//depth,width,lenght of data
+        double[][][] convolutionArray = new double[filterValue[layer].length][byteArray[0].length-2][byteArray[0][0].length-2];
 
         for (int i = 0; i < filterValue[layer].length; i++) {
             for (int j = 0; j < byteArray[0].length-2; j++) {
@@ -55,8 +55,8 @@ public class ConvolutionalNeuralNet {//convolution > RELU > Max pooling > neural
         return convolutionArray;
     }
 
-    public float[][][] convolution(float[][][] byteArray, int layer, int strideSize){//depth,width,lenght of data
-        float[][][] convolutionArray = new float[filterValue[layer].length][byteArray[0].length-2][byteArray[0][0].length-2];
+    public double[][][] convolution(double[][][] byteArray, int layer, int strideSize){//depth,width,lenght of data
+        double[][][] convolutionArray = new double[filterValue[layer].length][byteArray[0].length-2][byteArray[0][0].length-2];
 
         for (int i = 0; i < filterValue[layer].length; i++) {
             for (int j = 0; j < byteArray[0].length-2; j+=strideSize) {
@@ -68,8 +68,8 @@ public class ConvolutionalNeuralNet {//convolution > RELU > Max pooling > neural
         return convolutionArray;
     }
 
-    public float sumOfFilter(byte[][][] byteArray, int layer, int filterNumber, int yOffset, int xOffset){
-        float sum = 0;
+    public double sumOfFilter(byte[][][] byteArray, int layer, int filterNumber, int yOffset, int xOffset){
+        double sum = 0;
         for (int j = 0; j < filterValue[layer][filterNumber].length; j++) {//y
             for (int k = 0; k < filterValue[layer][filterNumber][j].length; k++) {//x
                 for (int i = 0; i < byteArray.length; i++) {
@@ -80,8 +80,8 @@ public class ConvolutionalNeuralNet {//convolution > RELU > Max pooling > neural
         return sum;
     }
 
-    public float sumOfFilter(float[][][] byteArray, int layer, int filterNumber, int yOffset, int xOffset){
-        float sum = 0;
+    public double sumOfFilter(double[][][] byteArray, int layer, int filterNumber, int yOffset, int xOffset){
+        double sum = 0;
         for (int j = 0; j < filterValue[layer][filterNumber].length; j++) {//y
             for (int k = 0; k < filterValue[layer][filterNumber][j].length; k++) {//x
                 for (int i = 0; i < byteArray.length; i++) {
@@ -92,8 +92,8 @@ public class ConvolutionalNeuralNet {//convolution > RELU > Max pooling > neural
         return sum;
     }
 
-    public float[][][] pooling(float[][][] byteArray, int poolingSize){
-        float[][][] poolingArray = new float[byteArray.length][byteArray[0].length/poolingSize][byteArray[0][0].length/poolingSize];
+    public double[][][] pooling(double[][][] byteArray, int poolingSize){
+        double[][][] poolingArray = new double[byteArray.length][byteArray[0].length/poolingSize][byteArray[0][0].length/poolingSize];
 
         for (int i = 0; i < poolingArray.length; i++) {
             for (int j = 0; j < poolingArray[i].length; j++) {
@@ -105,8 +105,8 @@ public class ConvolutionalNeuralNet {//convolution > RELU > Max pooling > neural
         return poolingArray;
     }
 
-    private float maxPool(float[][] values, int arrayOffsetY, int arrayOffsetX, int poolingSize){
-        float max = Float.NEGATIVE_INFINITY;
+    private double maxPool(double[][] values, int arrayOffsetY, int arrayOffsetX, int poolingSize){
+        double max = Double.NEGATIVE_INFINITY;
         for (int i = arrayOffsetY*poolingSize; i < arrayOffsetY*poolingSize+poolingSize; i++) {
             for (int j = arrayOffsetX*poolingSize; j < arrayOffsetX*poolingSize+poolingSize; j++) {
                 //System.out.print(values[i][j]+"     ");
@@ -124,18 +124,18 @@ public class ConvolutionalNeuralNet {//convolution > RELU > Max pooling > neural
             for (int j = 0; j < filterValue[i].length; j++) {
                 for (int k = 0; k < filterValue[i][j].length; k++) {
                     for (int l = 0; l < filterValue[i][j][k].length; l++) {
-                        filterValue[i][j][k][l] = ThreadLocalRandom.current().nextFloat();
+                        filterValue[i][j][k][l] = ThreadLocalRandom.current().nextDouble();
                     }
                 }
             }
         }
     }
 
-    /*public float[][] pooling(float[][] values, int poolingSize, int poolingStep){
-        float[][] poolingArray = new float[values.length/poolingSize][];
+    /*public double[][] pooling(double[][] values, int poolingSize, int poolingStep){
+        double[][] poolingArray = new double[values.length/poolingSize][];
 
         for (int i = 0; i < poolingArray.length; i++) {
-            float[] temp = new float[values[i].length/2];
+            double[] temp = new double[values[i].length/2];
             poolingArray[i] = temp;
         }
 
@@ -147,8 +147,8 @@ public class ConvolutionalNeuralNet {//convolution > RELU > Max pooling > neural
         return poolingArray;
     }
 
-    private float maxPool(float[][] values, int arrayOffsetY, int arrayOffsetX, int poolingSize){
-        float max = Float.NEGATIVE_INFINITY;
+    private double maxPool(double[][] values, int arrayOffsetY, int arrayOffsetX, int poolingSize){
+        double max = double.NEGATIVE_INFINITY;
         for (int i = arrayOffsetY*poolingSize; i < arrayOffsetY*poolingSize+poolingSize; i++) {
             for (int j = arrayOffsetX*poolingSize; j < arrayOffsetX*poolingSize+poolingSize; j++) {
                 //System.out.print(values[i][j]+"     ");
